+++ 4-func/sim/asm.cpp
@@ -7,7 +7,7 @@ assembler::assembler(std::istream &i) : i(i) {}
 
 void assembler::assembler_writer::write(const std::string &val, label) {
   fixup_labels[mem] = val;
-  ram_io::write(uint32_t{0}, label{});
+  ram_io::write(uintptr_t{0}, label{});
 }
 
 void assembler::assembler_writer::write(int val, reg) {
@@ -54,7 +54,7 @@ std::vector<unsigned char> assembler::run() && {
 
   for (auto &&[at, label] : writer.fixup_labels) {
     writer.mem = at;
-    writer.ram_io::write((uint32_t)label2offset[label]);
+    writer.ram_io::write(label2offset[label]);
   }
 
   return std::move(mem);
